scraper/requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
schedule==1.2.2
beautifulsoup4==4.13.3
lxml==5.3.1